Bring ReactLocalization back to encapsulate bundles and related methods
