+++ fluent-react/package.json
@@ -56,6 +56,7 @@
   },
   "devDependencies": {
     "@babel/plugin-proposal-async-generator-functions": "^7.2.0",
+    "@babel/plugin-proposal-class-properties": "^7.7.4",
     "@babel/preset-env": "^7.5.5",
     "@babel/preset-react": "7.0.0",
     "babel-jest": "^24.8.0",
