+++ fluent-react/src/context.js
@@ -1,3 +1,4 @@
 import { createContext } from "react";
+import ReactLocalization from "./localization";
 
-export default createContext({getBundle: null, parseMarkup: null});
+export default createContext(new ReactLocalization([]));
